+++ app/gui/controller/double-representation/src/text.rs
@@ -3,8 +3,7 @@
 use crate::prelude::*;
 
 use ast::IdMap;
-use enso_data::text::Size;
-use enso_data::text::Span;
+use enso_text::unit::*;
 
 
 
@@ -15,7 +14,7 @@ use enso_data::text::Span;
 /// Update IdMap to reflect the recent code change.
 pub fn apply_code_change_to_id_map(
     id_map: &mut IdMap,
-    change: &enso_data::text::TextChange,
+    change: &enso_text::text::Change<Bytes, String>,
     code: &str,
 ) {
     // TODO [mwu]
@@ -26,35 +25,33 @@ pub fn apply_code_change_to_id_map(
     //   API. Because of such expected rewrite and deeper restructuring, we don't really want to
     //   spend much time on refactoring this function right now, even if it could be made nicer.
 
-
-    let removed = change.replaced_span();
-    let inserted = change.inserted.as_str();
-    let new_code = change.applied(code);
+    let removed = &change.range.clone();
+    let inserted = change.text.as_str();
+    let new_code = change.applied(code).unwrap_or_else(|_| code.to_owned());
     let non_white = |c: char| !c.is_whitespace();
     let logger = enso_logger::DefaultWarningLogger::new("apply_code_change_to_id_map");
     let vector = &mut id_map.vec;
-    let inserted_size = Size::from_text(inserted);
+    let inserted_size: Bytes = inserted.len().into();
 
     info!(logger, "Old code:\n```\n{code}\n```");
     info!(logger, "New code:\n```\n{new_code}\n```");
     info!(logger, "Updating the ID map with the following text edit: {change:?}.");
 
     // Remove all entries fully covered by the removed span.
-    vector.drain_filter(|(span, _)| removed.contains_span(span));
+    vector.drain_filter(|(range, _)| removed.contains_range(range));
 
     // If the edited section ends up being the trailing part of AST node, how many bytes should be
     // trimmed from the id. Precalculated, as is constant in the loop below.
-    let to_trim_back = {
+    let to_trim_back: Bytes = {
         let last_non_white = inserted.rfind(non_white);
         let inserted_len = || inserted.len();
         let length_to_last_non_white = |index| inserted.len() - index - 1;
-        Size::new(last_non_white.map_or_else(inserted_len, length_to_last_non_white))
+        last_non_white.map_or_else(inserted_len, length_to_last_non_white).into()
     };
     // As above but for the front side.
-    let to_trim_front = {
+    let to_trim_front: Bytes = {
         let first_non_white = inserted.find(non_white);
-        let ret = first_non_white.unwrap_or_else(|| inserted.len());
-        Size::new(ret)
+        first_non_white.unwrap_or_else(|| inserted.len()).into()
     };
 
     let inserted_non_white = inserted.chars().any(non_white);
@@ -65,95 +62,98 @@ pub fn apply_code_change_to_id_map(
     // This is needed for edits like: `foo f` => `foo` — the earlier `foo` in `foo f` also has a
     // id map entry, however we want it to be consistently shadowed by the id from the whole App
     // expression.
-    let mut preferred: HashMap<Span, ast::Id> = default();
+    let mut preferred: HashMap<enso_text::Range<Bytes>, ast::Id> = default();
 
-    for (span, id) in vector.iter_mut() {
-        // These
+    for (range, id) in vector.iter_mut() {
         let mut trim_front = false;
         let mut trim_back = false;
-        let initial_span = *span;
-        info!(logger, "Processing @{span}: `{&code[*span]}`.");
-        if span.index > removed.end() {
+        let initial_range = *range;
+        info!(logger, "Processing @{range}: `{&code[*range]}`.");
+        if range.start > removed.end {
             debug!(logger, "Node after the edited region.");
             // AST node starts after edited region — it will be simply shifted.
-            let code_between = &code[Span::from(removed.end()..span.index)];
-            span.move_left(removed.size);
-            span.move_right(inserted_size);
+            let between_range: enso_text::Range<_> = (removed.end..range.start).into();
+            let code_between = &code[between_range];
+            *range = range.moved_left(removed.size()).moved_right(inserted_size);
 
             // If there are only spaces between current AST symbol and insertion, extend the symbol.
             // This is for cases like line with `foo ` being changed into `foo j`.
             debug!(logger, "Between: `{code_between}`.");
             if all_spaces(code_between) && inserted_non_white {
                 debug!(logger, "Will extend the node leftwards.");
-                span.extend_left(inserted_size);
-                span.extend_left(Size::from_text(code_between));
+                range.start -= inserted_size + between_range.size();
                 trim_front = true;
             }
-        } else if span.index >= removed.index {
-            // AST node starts inside the edited region. It doesn't end strictly inside it.
+        } else if range.start >= removed.start {
+            // AST node starts inside the edited region. It does not have to end inside it.
             debug!(logger, "Node overlapping with the end of the edited region.");
-            let removed_before = span.index - removed.index;
-            span.move_left(removed_before);
-            span.shrink_right(removed.size - removed_before);
-            span.extend_right(inserted_size);
+            let removed_before = range.start - removed.start;
+            *range = range.moved_left(removed_before);
+            range.end -= removed.size() - removed_before;
+            range.end += inserted_size;
             trim_front = true;
-        } else if span.end() >= removed.index {
+        } else if range.end >= removed.start {
             // AST node starts before the edited region and reaches (or possibly goes past) its end.
             debug!(logger, "Node overlapping with the beginning of the edited region.");
-            if span.end() <= removed.end() {
+            if range.end <= removed.end {
                 trim_back = true;
             }
-            let removed_chars = (span.end() - removed.index).min(removed.size);
-            span.shrink_right(removed_chars);
-            span.extend_right(inserted_size);
+            let removed_chars = (range.end - removed.start).min(removed.size());
+            range.end -= removed_chars;
+            range.end += inserted_size;
         } else {
             debug!(logger, "Node before the edited region.");
             // If there are only spaces between current AST symbol and insertion, extend the symbol.
             // This is for cases like line with `foo ` being changed into `foo j`.
-            let between = &code[Span::from(span.end()..removed.index)];
+            let between_range: enso_text::Range<_> = (range.end..removed.start).into();
+            let between = &code[between_range];
             if all_spaces(between) && inserted_non_white {
                 debug!(logger, "Will extend ");
-                span.size += Size::from_text(between) + inserted_size;
+                range.end += between_range.size() + inserted_size;
                 trim_back = true;
             }
         }
 
-        if trim_front && to_trim_front.non_empty() {
-            span.shrink_left(to_trim_front);
-            debug!(logger, "Trimming front {to_trim_front} bytes.");
+        if trim_front && to_trim_front > 0.bytes() {
+            range.start += to_trim_front;
+            debug!(logger, "Trimming front {to_trim_front.as_usize()} chars.");
         }
 
         if trim_back {
-            if to_trim_back.non_empty() {
-                span.shrink_right(to_trim_back);
-                debug!(logger, "Trimming back {to_trim_back} bytes.");
+            if to_trim_back > 0.bytes() {
+                range.end += -to_trim_back;
+                debug!(logger, "Trimming back {to_trim_back.as_usize()} chars.");
             }
-            let new_repr = &new_code[*span];
+            let new_repr = &new_code[*range];
             // Trim trailing spaces
-            let spaces = spaces_size(new_repr.chars().rev());
-            if spaces.non_empty() {
-                debug!(logger, "Additionally trimming {spaces} trailing spaces.");
+            let space_count = spaces_size(new_repr.chars().rev());
+            let spaces_len: Bytes = (space_count.as_usize() * ' '.len_utf8()).into();
+            if spaces_len > 0.bytes() {
+                debug!(logger, "Additionally trimming {space_count.as_usize()} trailing spaces.");
                 debug!(logger, "The would-be code: `{new_repr}`.");
-                span.shrink_right(spaces);
+                range.end -= spaces_len;
             }
         }
 
         // If we edited front or end of an AST node, its extended (or shrunk) span will be
         // preferred.
         if trim_front || trim_back {
-            preferred.insert(*span, *id);
+            preferred.insert(*range, *id);
         }
 
-        info!(
-            logger,
-            "Processing for id {id}: {initial_span} ->\t{span}.\n\
-        Code: `{&code[initial_span]}` => `{&new_code[*span]}`"
-        );
+        info!(logger, || {
+            let old_fragment = &code[initial_range];
+            let new_fragment = &new_code[*range];
+            iformat!(
+                "Processing for id {id}: {initial_range} ->\t{range}.\n
+                Code: `{old_fragment}` => `{new_fragment}`"
+            )
+        });
     }
 
     // If non-preferred entry collides with the preferred one, remove the former.
-    vector.drain_filter(|(span, id)| {
-        preferred.get(span).map(|preferred_id| id != preferred_id).unwrap_or(false)
+    vector.drain_filter(|(range, id)| {
+        preferred.get(range).map(|preferred_id| id != preferred_id).unwrap_or(false)
     });
 }
 
@@ -163,9 +163,9 @@ pub fn apply_code_change_to_id_map(
 // === Helpers ===
 // ===============
 
-/// Returns the byte length of leading space characters sequence.
-fn spaces_size(itr: impl Iterator<Item = char>) -> Size {
-    Size::new(itr.take_while(|c| *c == ' ').fold(0, |acc, _| acc + 1))
+/// Returns the chars count of leading space characters sequence.
+fn spaces_size(itr: impl Iterator<Item = char>) -> Chars {
+    itr.take_while(|c| *c == ' ').fold(0, |acc, _| acc + 1).into()
 }
 
 /// Checks if the given string slice contains only space charactesr.
@@ -186,8 +186,6 @@ mod test {
     use crate::module;
 
     use ast::HasIdMap;
-    use enso_data::text::Index;
-    use enso_data::text::TextChange;
     use enso_prelude::default;
     use parser::Parser;
     use uuid::Uuid;
@@ -195,11 +193,12 @@ mod test {
     /// A sample text edit used to test "text api" properties.
     ///
     /// See `from_markdown` constructor function for helper markdown description.
+    #[derive(Debug)]
     struct Case {
         /// The initial enso program code.
         pub code:   String,
         /// The edit made to the initial code.
-        pub change: TextChange,
+        pub change: enso_text::Change<Bytes, String>,
     }
 
     impl Case {
@@ -227,11 +226,9 @@ mod test {
                     let inserted_code = insertion.map_or("", |insertion| {
                         &marked_code[insertion + INSERTION.len_utf8()..end]
                     });
-                    let removed_span = Range {
-                        start: Index::new(start),
-                        end:   Index::new(erased_finish - START.len_utf8()),
-                    };
-                    let change = TextChange::replace(removed_span, inserted_code.to_string());
+                    let range_end = (erased_finish - START.len_utf8()).into();
+                    let range = enso_text::Range::new(start.into(), range_end);
+                    let change = enso_text::Change { range, text: inserted_code.to_string() };
                     Case { code, change }
                 }
                 _ => panic!("Invalid markdown in the marked code: {}.", marked_code),
@@ -240,7 +237,7 @@ mod test {
 
         /// Code after applying the change
         fn resulting_code(&self) -> String {
-            self.change.applied(&self.code)
+            self.change.applied(&self.code).expect("Change removed range out of bounds")
         }
 
         /// Checks if the text operation described by this case keeps the node IDs intact.
@@ -254,7 +251,17 @@ mod test {
             let code2 = self.resulting_code();
 
             let ast2 = parser.parse_module(&code2, id_map.clone()).unwrap();
-            assert_same_node_ids(&ast1, &ast2);
+            self.assert_same_node_ids(&ast1, &ast2);
+        }
+
+        /// Checks that both module AST contain `main` function that has the same sequence of node
+        /// IDs, as described by the `main_nodes` function.
+        fn assert_same_node_ids(&self, ast1: &ast::known::Module, ast2: &ast::known::Module) {
+            let ids1 = main_nodes(ast1);
+            let ids2 = main_nodes(ast2);
+            DEBUG!("IDs1: {ids1:?}");
+            DEBUG!("IDs2: {ids2:?}");
+            assert_eq!(ids1, ids2, "Node ids mismatch in {:?}", self);
         }
     }
 
@@ -279,28 +286,18 @@ mod test {
         nodes.into_iter().map(|node| node.id()).collect()
     }
 
-    /// Checks that both module AST contain `main` function that has the same sequence of node IDs,
-    /// as described by the `main_nodes` function.
-    fn assert_same_node_ids(ast1: &ast::known::Module, ast2: &ast::known::Module) {
-        let ids1 = main_nodes(ast1);
-        let ids2 = main_nodes(ast2);
-        DEBUG!("IDs1: {ids1:?}");
-        DEBUG!("IDs2: {ids2:?}");
-        assert_eq!(ids1, ids2);
-    }
-
     #[test]
     fn test_case_markdown() {
         let case = Case::from_markdown("foo«aa⎀bb»c");
         assert_eq!(case.code, "fooaac");
-        assert_eq!(case.change.inserted, "bb");
-        assert_eq!(case.change.replaced, Index::new(3)..Index::new(5));
+        assert_eq!(case.change.text, "bb");
+        assert_eq!(case.change.range, 3.bytes()..5.bytes());
         assert_eq!(case.resulting_code(), "foobbc");
 
         let case = Case::from_markdown("foo«aa»c");
         assert_eq!(case.code, "fooaac");
-        assert_eq!(case.change.inserted, "");
-        assert_eq!(case.change.replaced, Index::new(3)..Index::new(5));
+        assert_eq!(case.change.text, "");
+        assert_eq!(case.change.range, 3.bytes()..5.bytes());
         assert_eq!(case.resulting_code(), "fooc");
     }