+++ engine/language-server/src/main/scala/org/enso/languageserver/libraries/LibraryInstallerConfig.scala
@@ -1,7 +1,10 @@
 package org.enso.languageserver.libraries
 
-import org.enso.distribution.{DistributionManager, LanguageHome}
 import org.enso.distribution.locking.ResourceManager
+import org.enso.distribution.{DistributionManager, LanguageHome}
+import org.enso.libraryupload.DependencyExtractor
+
+import java.io.File
 
 /** Gathers configuration needed by the library installer used in the
   * `library/preinstall` endpoint.
@@ -9,9 +12,11 @@ import org.enso.distribution.locking.ResourceManager
   * @param distributionManager the distribution manager
   * @param resourceManager a resource manager instance
   * @param languageHome language home, if detected / applicable
+  * @param dependencyExtractor a dependency extractor
   */
 case class LibraryInstallerConfig(
   distributionManager: DistributionManager,
   resourceManager: ResourceManager,
-  languageHome: Option[LanguageHome]
+  languageHome: Option[LanguageHome],
+  dependencyExtractor: DependencyExtractor[File]
 )