+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/json/JsonConnectionController.scala
@@ -513,10 +513,8 @@ class JsonConnectionController(
         .props(requestTimeout, libraryConfig.localLibraryManager),
       LibraryGetMetadata -> LibraryGetMetadataHandler
         .props(requestTimeout, libraryConfig.localLibraryManager),
-      LibraryPreinstall -> LibraryPreinstallHandler.props(
-        libraryConfig.editionReferenceResolver,
-        libraryConfig.installerConfig
-      ),
+      LibraryPreinstall -> LibraryPreinstallHandler
+        .props(libraryConfig.editionReferenceResolver, libraryConfig),
       LibraryPublish -> LibraryPublishHandler
         .props(requestTimeout, libraryConfig.localLibraryManager),
       LibrarySetMetadata -> LibrarySetMetadataHandler