+++ lib/rust/text/src/range.rs
@@ -3,8 +3,8 @@
 
 use crate::prelude::*;
 
-use super::rope;
-use super::unit::*;
+use crate::rope;
+use crate::unit::*;
 
 
 
@@ -16,7 +16,7 @@ use super::unit::*;
 ///
 /// Unlike `std::ops::Range`, this type implements `Copy`, and contains text-related trait
 /// implementations.
-#[derive(Clone, Copy, Default, PartialEq, Eq)]
+#[derive(Clone, Copy, Default, PartialEq, Eq, Hash)]
 #[allow(missing_docs)]
 pub struct Range<T> {
     pub start: T,
@@ -49,10 +49,22 @@ impl<T> Range<T> {
         Self { start, end }
     }
 
+    /// Return new range with the `offset` subtracted from both ends.
+    pub fn moved_left(&self, offset: T) -> Self
+    where T: Clone + Sub<T, Output = T> {
+        Self { start: self.start.clone() - offset.clone(), end: self.end.clone() - offset }
+    }
+
+    /// Return new range with the `offset` added to both ends.
+    pub fn moved_right(&self, offset: T) -> Self
+    where T: Clone + Add<T, Output = T> {
+        Self { start: self.start.clone() + offset.clone(), end: self.end.clone() + offset }
+    }
+
     /// Map both values with the provided function.
-    pub fn map(&self, f: impl Fn(T) -> T) -> Self
+    pub fn map<U>(&self, f: impl Fn(T) -> U) -> Range<U>
     where T: Clone {
-        self.with_start(f(self.start.clone())).with_end(f(self.end.clone()))
+        Range { start: f(self.start.clone()), end: f(self.end.clone()) }
     }
 
     /// Map the start value with the provided function.
@@ -65,6 +77,20 @@ impl<T> Range<T> {
     pub fn map_end(&self, f: impl FnOnce(T) -> T) -> Self
     where T: Clone {
         self.with_end(f(self.end.clone()))
+    }
+
+    /// Check if the range contains the given value.
+    pub fn contains<U>(&self, value: &U) -> bool
+    where
+        T: PartialOrd<U>,
+        U: PartialOrd<T>, {
+        value >= &self.start && value < &self.end
+    }
+
+    /// Check if the range contains all values from `other` range.
+    pub fn contains_range(&self, other: &Range<T>) -> bool
+    where T: PartialOrd {
+        self.start <= other.start && self.end >= other.end
     }
 }
 
@@ -93,10 +119,16 @@ impl<T: Debug> Debug for Range<T> {
     }
 }
 
-impl<T> From<std::ops::Range<T>> for Range<T> {
-    fn from(range: std::ops::Range<T>) -> Range<T> {
+impl<T, U: Into<T>> From<std::ops::Range<U>> for Range<T> {
+    fn from(range: std::ops::Range<U>) -> Range<T> {
         let std::ops::Range { start, end } = range;
-        Range { start, end }
+        Range { start: start.into(), end: end.into() }
+    }
+}
+
+impl<T: PartialEq<T>> PartialEq<std::ops::Range<T>> for Range<T> {
+    fn eq(&self, other: &std::ops::Range<T>) -> bool {
+        (&self.start, &self.end) == (&other.start, &other.end)
     }
 }
 
@@ -118,6 +150,24 @@ impl From<RangeInclusive<Bytes>> for Range<Bytes> {
 impl From<RangeToInclusive<Bytes>> for Range<Bytes> {
     fn from(range: RangeToInclusive<Bytes>) -> Range<Bytes> {
         Range::new(0.bytes(), range.end.saturating_add(1.bytes()))
+    }
+}
+
+impl Index<Range<Bytes>> for str {
+    type Output = str;
+
+    fn index(&self, index: Range<Bytes>) -> &Self::Output {
+        let start = index.start.as_usize();
+        let end = index.end.as_usize();
+        &self[start..end]
+    }
+}
+
+impl Index<Range<Bytes>> for String {
+    type Output = str;
+
+    fn index(&self, index: Range<Bytes>) -> &Self::Output {
+        &self.as_str()[index]
     }
 }